requirements.txt
NumPy==2.1.3
Pandas==2.2.3
torch>=2.1.0
notebook==7.3.2
torch>=2.1.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
spacy>=3.5.0
easyocr opencv-python
spacy
pywin32